# gymScore-backend/docs/swagger.yaml
basePath: /
definitions:
  models.AceitarAmizadeRequest:
    properties:
      id_amigo:
        type: integer
      id_usuario:
        type: integer
    required:
    - id_amigo
    - id_usuario
    type: object
...
  models.AdicionarAmigoRequest:
    properties:
      id_amigo:
        type: integer
      id_usuario:
        type: integer
    required:
    - id_amigo
    - id_usuario
    type: object
...
  models.RemoverAmigoRequest:
    properties:
      id_amigo:
        type: integer
      id_usuario:
        type: integer
    required:
    - id_amigo
    - id_usuario
    type: object
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API para o sistema GymScore de desafios e treinos.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GymScore API
  version: "1.0"
paths:
  /api/amigos/{id}:
    get:
      description: Retorna a lista de amigos aceitos de um usuário
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Listar amigos de um usuário
      tags:
      - amigos
  /api/amigos/aceitar:
    post:
      consumes:
      - application/json
      description: Aceita uma solicitação de amizade pendente
      parameters:
      - description: IDs dos usuários
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AceitarAmizadeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Aceitar solicitação de amizade
      tags:
      - amigos
  /api/amigos/adicionar:
    post:
      consumes:
      - application/json
      description: Envia uma solicitação de amizade para outro usuário
      parameters:
      - description: IDs dos usuários
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AdicionarAmigoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Adicionar amigo
      tags:
      - amigos
  /api/amigos/remover:
    post:
      consumes:
      - application/json
      description: Remove o vínculo de amizade entre dois usuários
      parameters:
      - description: IDs dos usuários
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RemoverAmigoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Remover amigo
      tags:
      - amigos
...
securityDefinitions:
  BearerAuth:
    description: 'Digite "Bearer " seguido do seu token JWT. Exemplo: "Bearer eyJhbGciOiJIUzI1..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"